+++ Muscollo/tests/testSingleMuscleDeGrooteFregly2016.cpp
@@ -124,7 +124,7 @@ solveForTrajectoryGSO() {
                                 "joint/height/speed"});
     const auto& position = ocpSolution.getDependentColumn("position");
     const auto& speed = ocpSolution.getDependentColumn("speed");
-    for (size_t iRow = 0; iRow < ocpSolution.getNumRows(); ++iRow) {
+    for (int iRow = 0; iRow < (int)ocpSolution.getNumRows(); ++iRow) {
         SimTK::RowVector row(2);
         row[0] = position[iRow];
         row[1] = speed[iRow];
@@ -296,7 +296,7 @@ solveForTrajectoryINDYGO() {
                                 "joint/height/speed"});
     const auto& position = ocpSolution.getDependentColumn("position");
     const auto& speed = ocpSolution.getDependentColumn("speed");
-    for (size_t iRow = 0; iRow < ocpSolution.getNumRows(); ++iRow) {
+    for (int iRow = 0; iRow < (int)ocpSolution.getNumRows(); ++iRow) {
         SimTK::RowVector row(2);
         row[0] = position[iRow];
         row[1] = speed[iRow];
